# cpu: matmul: enable optimized 8-bit gemm for s390x