@quickwritereader 2023-02-09
@igorsafo I tested them using gtests/test_gemm_s8u8s32, gtests/test_gemm_s8s8s32. I have not used onednn bench yet
 for performance checks, I used it only in my local work and also tested them
https://github.com/quickwritereader/GemmLowZ15
I tried to use the u8u8-packed vec_sum4 kernel as a base as well. but I faced several problems there
firstly without having subrange limitations I was getting a worse performance.
plus to convert it to s8u8 or s8s8 kernels one should add a shift to convert them into uint8_t  and also the same trick with ao, bo shifts (https://github.com/oneapi-src/oneDNN/blob/master/src/cpu/x64/gemm/gemm_driver.cpp#L412-L481)
the s8s8 and u8u8 approaches give a low performance.
in the end, I chose this one as it was better performance and maintenance-wise.
the one drawback for all above is that I have to create a C buffer for correctness as scaling C with beta beforehand gives +1 difference.
I tried all below:
```
//could be used for all as it easily promoted, 2 ops,  assuming it does twice it means **4 ops**
inline vec_type_t<int32_t> multiplyAdd(vec_type_t<int16_t> va,
        vec_type_t<int16_t> vb, vec_type_t<int32_t> vc) { 
    // 2 ops  2 moad
    auto a = va.vec();
    auto b = vb.vec();
    auto c = vc.vec();
    c = vec_moadd(a, b, c);
    c = vec_meadd(a, b, c); 
    return vec_type_t<int32_t> {c};
}

//u8u8   . for k 2  it was possible to gain a little speed up using vec_sum4's second value by permutations, but still, it was **5 ops** 
inline vec_type_t<uint32_t> multiplySum4(vec_type_t<uint8_t> va, vec_type_t<uint8_t> vb, vec_type_t<uint32_t> vc ) {
    // 6 ops  2 mul 2 vec_sum 2 addition
    const vuint16 vz16 = {};
    const auto a = va.vec();
    const auto b = vb.vec();
    auto c = vc.vec();
    auto reso = vec_mulo(a, b);
    auto rese = vec_mule(a, b);

    c= c + vec_sum4(reso, vz16) + vec_sum4(rese, vz16); 
    return vec_type_t<uint32_t> {c};
}

//s8s8 or u8u8 kernel , still needs additional computation for s8u8 **6ops**
inline vec_type_t<uint32_t> multiplyAdd(vec_type_t<uint8_t> va,
        vec_type_t<uint8_t> vb, vec_type_t<uint32_t> vc) { 
    // 6 ops
    vuint8 a = va.vec();
    vuint8 b = vb.vec();
    auto c = vc.vec();
    const vuint16 vone16 = {1, 1, 1, 1, 1, 1, 1, 1};
    vuint16 reso = vec_mulo(a, b);
    vuint16 rese = vec_mule(a, b);
    c = vec_moadd(reso, vone16, c);
    c = vec_meadd(reso, vone16, c);
    c = vec_moadd(rese, vone16, c);
    c = vec_meadd(rese, vone16, c); 
    return vec_type_t<uint32_t> {c};
}


 //u8u8   one of the matrices should be in the range [-64;63] [0;127]   **4ops**
inline vec_type_t<uint32_t> multiplySum4Low(vec_type_t<uint8_t> va, vec_type_t<uint8_t> vb, vec_type_t<uint32_t> vc) {
    // 4 ops  2 mul 1 vec_sum 1 addition
    const vuint16 vz16 = {};
    const auto a = va.vec();
    const auto b = vb.vec();
    auto c = vc.vec();
    vuint16 d = vec_mulo(a, b);
    vuint16 e = vec_meadd(a, b, d);
    c= c + vec_sum4(e, vz16);  
    return vec_type_t<uint32_t> {c};
}
```