@quickwritereader 2023-02-09
generated assembly in the s16s16s32 kernel: https://godbolt.org/z/G8a6Ecq5n