@quickwritereader 2023-02-11
[test_gemm_s8s8s32.txt and test_gemm_u8s8s32.txt](https://gist.github.com/quickwritereader/e10a5ca463a9e84b86bb41393b8ee0ed)