@quickwritereader 2023-02-11
` ./tests/benchdnn/benchdnn --matmul --mode=P --dt=s8:s8:s32 1000x1000:1000x1000`

New:
 export OMP_NUM_THREADS=1
 ```
Output template: perf,%engine%,%impl%,%name%,%prb%,%Gops%,%-time%,%-Gflops%,%0time%,%0Gflops%
perf,cpu,gemm:jit,,--matmul --mode=P --dt=s8:s8:s32 1000x1000:1000x1000,2,32.8674,60.8505,33.6525,59.431
tests:1 passed:1 skipped:0 mistrusted:0 unimplemented:0 invalid_arguments:0 failed:0 listed:0
total perf: min(ms):32.8674 avg(ms):33.6525


```
export OMP_NUM_THREADS=2

```
 Output template: perf,%engine%,%impl%,%name%,%prb%,%Gops%,%-time%,%-Gflops%,%0time%,%0Gflops%
perf,cpu,gemm:jit,,--matmul --mode=P --dt=s8:s8:s32 1000x1000:1000x1000,2,16.5764,120.653,17.2736,115.784
tests:1 passed:1 skipped:0 mistrusted:0 unimplemented:0 invalid_arguments:0 failed:0 listed:0
total perf: min(ms):16.5764 avg(ms):17.2736

```

Old one:

```
 export OMP_NUM_THREADS=1
```
```
Output template: perf,%engine%,%impl%,%name%,%prb%,%Gops%,%-time%,%-Gflops%,%0time%,%0Gflops%
perf,cpu,gemm:jit,,--matmul --mode=P --dt=s8:s8:s32 1000x1000:1000x1000,2,120.126,16.6492,154.963,12.9063
tests:1 passed:1 skipped:0 mistrusted:0 unimplemented:0 invalid_arguments:0 failed:0 listed:0
total perf: min(ms):120.126 avg(ms):154.963
```

 export OMP_NUM_THREADS=2
```
Output template: perf,%engine%,%impl%,%name%,%prb%,%Gops%,%-time%,%-Gflops%,%0time%,%0Gflops%
perf,cpu,gemm:jit,,--matmul --mode=P --dt=s8:s8:s32 1000x1000:1000x1000,2,55.085,36.3076,57.5478,34.7537
tests:1 passed:1 skipped:0 mistrusted:0 unimplemented:0 invalid_arguments:0 failed:0 listed:0
total perf: min(ms):55.085 avg(ms):57.5478
```