@quickwritereader 2023-02-11
from the bench, it shows that the provided kernel performs 115-120Gops, but the old one does 34-36Gops

notes:
the reason why it does not perform more than expected is because of the current limitations of the architecture. I explained above why I had to go with the s16s16s32 `kernel which yields performance similar to or better than float32 optimized gemm.  the other good performing one was the limited range [0;127] uint8 kernel, but still, it was not good enough to choose it, assuming further additional computations had to be done for int8->uint8 and ob ao shifts.