@quickwritereader 2023-02-14
well, nowadays, machine learning and high performance related tasks are applied everywhere. And onednn is used by many projects.
and their recent ibm z16 comes even with on chip AI .