@quickwritereader 2023-02-15
offtopic:
@igorsafo, @mgouicem 
team, I think I can provide floating gemm as well. before that I will check if there is 1.2-1.3x speedup at least. But afaik onednn can use just blas gemm as well.
besides that, if it is ok to try to optimize other operations. I see that most of codes are jit based. thats why I am asking.
if its ok to provide nonjited version of those operations then for convolution, could you tell me what approach is used in onednn so that I give it a try. I know that its possible to do it by img2col + gemm calls. but there are other options as well.
thanks.